Guard for the case where payee or sender does not exist.
If the payee or (more likely) sender is not in the CSV file, running alist-get
on it without a default value will yield nil. When passed to string-empty-p, the
result is nil, because the symbol name of nil is not an empty string. This could
lead to nil appearing as the title of an entry. By having alist-get return the
empty string as default value, this is avoided.